refactor: remove separated login page
